+++ corebehrt/ozempic/utils/calculations.py
