+++ corebehrt/ozempic/utils/definitions.py
